did sorting

+++ Include Files/Metrics.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 
 #define KILOMETER 100
 #define HECTOMETER 10
@@ -30,8 +31,8 @@ typedef signed int int32;
 typedef unsigned int uint;
 //#define UINT_MAX 4294967295
 
-typedef long long int64;
-typedef unsigned long long uint64;
+typedef int64_t  int64;
+typedef uint64_t  uint64;
 
 typedef wchar_t wchar;
